+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 ### is.codion.framework.model
 - DefaultEntityQueryModel.conditionChanged() bug fixed, no longer active when limit, orderBy or attributes have changed.
 - EntityQueryModel.resetConditionChanged() removed.
+- EntitySearchModel.Builder.condition() added.
 ### is.codion.swing.framework.ui
 - EntityTablePanel.SHOW_REFRESH_PROGRESS_BAR renamed REFRESH_PROGRESS_BAR, now true by default.
 - EntitySearchField bug fixed, no longer updates the background color when disabled.